Add files via upload
data_bou.mat and data_boh.mat are two two different data files based on the same simulation model but different healthy conditions. data_bou.mat is an outer_ring dataset and data_boh.mat is the healthy dataset.  For each dataset， it consists of 4000 univariate time series sample with length of 512. All of the data is actually not normalized. The data preprocessing can be seen in the attached Jupyter file.